added deletion functionnality to admin

+++ src/main/java/spring/bricole/service/JobService.java
@@ -5,7 +5,6 @@ import spring.bricole.model.Job;
 import spring.bricole.repository.JobRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class JobService {
@@ -47,7 +46,21 @@ public class JobService {
     }
 
     // Delete job by ID
-    public void deleteJobById(Integer id) {
-        jobRepository.deleteById(id);
+    public boolean deleteJobById(Integer id) {
+        if(jobRepository.existsById(id)) {
+            jobRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    // delete job
+    public boolean deleteJobById(int id) {
+        if(jobRepository.existsById(id)){
+            jobRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
